Key Highlights
- Improved caching and selection mechanisms for miners in the validator/api_server.py file.
- Added new utility functions for fetching and handling miner-related data.
- Enhanced code readability and maintainability with better logging and structured updates.

+++ shared/veridex_protocol.py
@@ -50,7 +50,7 @@ class VericoreStatementResponse():
 @dataclass
 class VericoreMinerStatementResponse():
   miner_hotkey: str
-  miner_uid: float
+  miner_uid: int
   status: str
   vericore_responses: typing.List["VericoreStatementResponse"] = field(default_factory=list)
   speed_factor: float = 0
